python-examples/RAG + internet browsing/requirements.txt
beautifulsoup4==4.12.2
colorama==0.4.4
chromadb==0.4.18
langchain==0.0.347
openai==1.3.9
python-dotenv==1.0.0
streamlit_chat==0.1.1
streamlit==1.20.0
tiktoken==0.5.2
langchain==0.0.347
openai==1.3.9
pandas==2.1.4
streamlit==1.29.0
tqdm==4.66.1
html5lib==1.1
trafilatura==0.5.0
